pyproject.toml
[tool.poetry]
name = "sophos_artifacts"
version = "0.1.0"
description = ""
authors = ["ranguli <hello@joshmurphy.ca>"]

[tool.poetry.dependencies]
python = "^3.6"
OTXv2 = "^1.5.7"
python-decouple = "^3.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
